+++ src/main/java/de/bixilon/minosoft/data/entities/block/BlockActionEntity.kt
@@ -1,6 +1,6 @@
 /*
  * Minosoft
- * Copyright (C) 2021 Moritz Zwerger
+ * Copyright (C) 2020-2023 Moritz Zwerger
  *
  * This program is free software: you can redistribute it and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.
  *
@@ -14,5 +14,5 @@
 package de.bixilon.minosoft.data.entities.block
 
 interface BlockActionEntity {
-    fun setBlockActionData(data1: Byte, data2: Byte)
+    fun setBlockActionData(data1: Int, data2: Int)
 }
