+++ src/main/java/de/bixilon/minosoft/data/entities/block/redstone/piston/PistonBlockEntity.kt
@@ -22,7 +22,6 @@ import de.bixilon.minosoft.data.entities.block.BlockEntityFactory
 import de.bixilon.minosoft.data.registries.identified.Namespaces.minecraft
 import de.bixilon.minosoft.data.registries.identified.ResourceLocation
 import de.bixilon.minosoft.protocol.network.connection.play.PlayConnection
-import de.bixilon.minosoft.util.KUtil
 
 open class PistonBlockEntity(connection: PlayConnection) : BlockEntity(connection), BlockActionEntity {
     var state: PistonStates = PistonStates.PULL
@@ -30,7 +29,7 @@ open class PistonBlockEntity(connection: PlayConnection) : BlockEntity(connectio
     var direction: Directions = Directions.NORTH
         private set
 
-    override fun setBlockActionData(data1: Byte, data2: Byte) {
+    override fun setBlockActionData(data1: Int, data2: Int) {
         state = PistonStates[data1.toInt()]
         direction = Directions[data2.toInt()]
     }
